toml-ify the project. refactoring so that 'ufm' is a proper module
