--- requirements.txt
@@ -1,10 +1,0 @@
-matplotlib
-seaborn
-tqdm
-torch
-torchvision
-ipython
-ipykernel
-ipywidgets
-omegaconf
-numpy
